implement test

+++ services/userService.js
@@ -3,6 +3,8 @@ import Role from '../models/roleModel.js';
 import bcrypt from 'bcrypt';
 import jwt from 'jsonwebtoken';
 import validator from 'validator';
+import AppError from "../utils/AppError.js";
+import mongoose from "mongoose";
 
 const createToken = (user) => {
     return jwt.sign({ id: user._id }, process.env.JWT_SECRET, { expiresIn: '1h' });
@@ -16,24 +18,24 @@ const registerUserService = async ({
 }) => {
     try {
         if (!name || !email || !password) {
-            throw new Error("Name, email and password are required");
+            throw new AppError("Name, email and password are required", 400);
         }
         if (!validator.isEmail(email)) {
-            throw new Error("Invalid email format");
+            throw new AppError("Invalid email format", 400);
         }
         if (password.length < 6) {
-            throw new Error("Password must be at least 6 characters long");
+            throw new AppError("Password must be at least 6 characters long", 400);
         }
         const existingUser = await User.findOne({ email });
         if (existingUser) {
-            throw new Error("Email already in use");
+            throw new AppError("Email already in use", 400);
         }
         let userRole;
 
         if (role) {
             userRole = await Role.findById(role);
             if (!userRole) {
-                throw new Error("Invalid role provided");
+                throw new AppError("Invalid role provided", 400);
             }
         } else {
             userRole = await Role.findOne({ name: "user" });
@@ -56,7 +58,6 @@ const registerUserService = async ({
             token
         };
     } catch (error) {
-        // console.error("Error registering user:", error.message);
         throw error;
     }
 };
@@ -65,7 +66,7 @@ const updateUserService = async (userId, updateData) => {
     const { name, email, password, role } = updateData;
     const user = await User.findById(userId);
     if (!user) {
-        throw new Error("User not found");
+        throw new AppError("User not found", 404);
     }
     const payload = {};
     if (name) {
@@ -73,7 +74,7 @@ const updateUserService = async (userId, updateData) => {
     }
     if (email) {
         if (!validator.isEmail(email)) {
-            throw new Error("Invalid email format");
+            throw new AppError("Invalid email format", 400);
         }
         const existingUser = await User.findOne({
             email,
@@ -113,23 +114,31 @@ const updateUserService = async (userId, updateData) => {
 };
 
 const deleteUserService = async (userId) => {
+
+    if (!mongoose.Types.ObjectId.isValid(userId)) {
+        throw new AppError("Invalid user ID", 400);
+    }
+
     const user = await User.findById(userId);
     if (!user) {
-        throw new Error("User not found");
+        throw new AppError("User not found", 404);
     }
     await User.findByIdAndDelete(userId);
     return true;
 };
 
 const loginUserService = async ({ email, password }) => {
+    if (!email || !password) {
+        throw new AppError("Email and password are required", 400);
+
+    }
     const user = await User.findOne({ email }).populate("role");
-    console.log(user)
     if (!user) {
-        throw new Error("User not found.");
+        throw new AppError("User not found.", 401);
     }
     const isMatch = await bcrypt.compare(password, user.password);
     if (!isMatch) {
-        throw new Error("Invalid email or password");
+        throw new AppError("Invalid email or password", 401);
     }
     const token = createToken(user);
     return { user, token };
